Started work on a more visual db component
As the title says I've started work on a more visual way of traversing the database. This is in line with the idea of the app being made for use on a phone and a step towards a greater ux and ui since the current one is very "back end debug" looking.

Learned a lot about dom manipulation and data types already and am looking forward to getting this version of the database interface operational!